* fixed warning C4242: '=' : conversion from 'sector_t' to 'long', possible loss of data (256) ** add BUG_ON() if there is a possibility of overflow ** add forced Type Casting

+++ wdrbd9/wsk2.c
@@ -1292,7 +1292,7 @@ LONG NTAPI Receive(
 		KeWaitForSingleObject(&CompletionEvent, Executive, KernelMode, FALSE, NULL);
 		if (Irp->IoStatus.Information > 0) {
 			//WDRBD_INFO("rx canceled but rx data(%d) avaliable.\n", Irp->IoStatus.Information);
-			BytesReceived = Irp->IoStatus.Information;
+			BytesReceived = (LONG)Irp->IoStatus.Information;
 		}
 	}
